+++ src/main/java/subway/presentation/dto/request/CreationMiddleSectionRequest.java
@@ -1,4 +1,4 @@
-package subway.ui.dto.request;
+package subway.presentation.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
