+++ src/main/java/subway/presentation/dto/response/AddSectionResponse.java
@@ -1,4 +1,4 @@
-package subway.ui.dto.response;
+package subway.presentation.dto.response;
 
 import java.util.List;
 import java.util.stream.Collectors;
