+++ src/main/java/subway/presentation/dto/response/CreationStationResponse.java
@@ -1,4 +1,4 @@
-package subway.ui.dto.response;
+package subway.presentation.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
